changed this file

+++ add_function.py
@@ -1,8 +1,8 @@
 def add(x,y):
     return x+y
 
-x = float(input())
-y = float(input())
+x = int(input())
+y = int(input())
 
 result = add(x,y)
 print(result)
